refactor: corrections to comments on homework 2
