+++ src/app/App.tsx
@@ -1,7 +1,6 @@
 import '../shared/styles/index.css';
 import {useEffect} from 'react';
 import useTheme from '../shared/lib/theme/UseTheme.ts';
-import {Theme} from "../shared/lib/theme/ThemeContext.ts";
 import MainLayout from '../shared/layouts/MainLayout';
 import HomePage from '../pages/home';
 import LayoutHeader from '../widgets/LayoutHeader';
@@ -13,8 +12,7 @@ const App = () => {
   const {theme} = useTheme();
 
   useEffect(() => {
-    document.body.classList.remove(Theme.Light, Theme.Dark);
-    document.body.classList.add(theme);
+    document.documentElement.dataset.theme = `${theme}`;
   }, [theme]);
 
   return (
