+++ src/shared/lib/theme/ThemeContext.ts
@@ -1,7 +1,6 @@
 import {createContext} from 'react';
+import type {ThemeType} from '../../constants/constants.ts';
 
-
-type ThemeType = typeof Theme[keyof typeof Theme];
 
 type ThemeContextType = {
   theme?: ThemeType,
@@ -9,12 +8,7 @@ type ThemeContextType = {
 };
 
 
-const Theme = {
-  Dark: 'theme-dark',
-  Light: 'theme-light'
-} as const;
-
 const ThemeContext = createContext<ThemeContextType>({});
 
 
-export {type ThemeType, type ThemeContextType, Theme, ThemeContext};
+export {type ThemeContextType, ThemeContext};
